+++ addon/policies/exponential-backoff.js
@@ -1,5 +1,5 @@
 import { assert } from '@ember/debug';
-import DelayPolicy from "./delay";
+import DelayPolicy from './delay';
 
 function buildIntervals(multiplier, minDelay, maxDelay) {
   let currentInterval = minDelay;
@@ -8,7 +8,7 @@ function buildIntervals(multiplier, minDelay, maxDelay) {
   do {
     intervals.push(currentInterval);
     currentInterval = currentInterval * multiplier;
-  } while(currentInterval <= maxDelay);
+  } while (currentInterval <= maxDelay);
 
   intervals.push(maxDelay);
 
@@ -33,9 +33,18 @@ export default class ExponentialBackoffPolicy extends DelayPolicy {
    * @param {number} [options.maxDelay=500] Maximum delay duration (in milliseconds)
    */
   constructor({ multiplier = 2, minDelay = 50, maxDelay = 500, reasons = [] }) {
-    assert("The `multiplier` argument must be a Number", typeof multiplier === 'number');
-    assert("The `minDelay` argument must be a Number", typeof minDelay === 'number');
-    assert("The `maxDelay` argument must be a Number", typeof maxDelay === 'number');
+    assert(
+      'The `multiplier` argument must be a Number',
+      typeof multiplier === 'number'
+    );
+    assert(
+      'The `minDelay` argument must be a Number',
+      typeof minDelay === 'number'
+    );
+    assert(
+      'The `maxDelay` argument must be a Number',
+      typeof maxDelay === 'number'
+    );
 
     const delay = buildIntervals(multiplier, minDelay, maxDelay);
     super({ delay, reasons });