+++ tests/dummy/app/components/retryable-demo.js
@@ -1,91 +1,125 @@
-import defineModifier from 'ember-concurrency-retryable/define-modifier';
-defineModifier();
-
-/* eslint-disable ember/no-on-calls-in-components */
-/* eslint-disable ember/no-classic-components */
-
 // BEGIN-SNIPPET retryable-demo.js
-import Component from '@ember/component';
-import { on } from '@ember/object/evented';
-import {
-  task,
-  timeout
-} from 'ember-concurrency';
-import ExponentialBackoffPolicy from 'ember-concurrency-retryable/policies/exponential-backoff';
+import Component from '@glimmer/component';
+import { tracked } from '@glimmer/tracking';
+import { registerDestructor } from '@ember/destroyable';
+import { addListener, removeListener, sendEvent } from '@ember/object/events';
+import { task, timeout } from 'ember-concurrency';
+import { ExponentialBackoffPolicy } from 'ember-concurrency-retryable';
 
 const COLORS = ['blue', 'pink', 'darkred', 'red', 'teal'];
 const backoffPolicy = new ExponentialBackoffPolicy({
   multiplier: 1.5,
   minDelay: 30,
-  maxDelay: 400
+  maxDelay: 400,
 });
 
-export default Component.extend({
-  tagName: '',
-
-  ajaxTask: task(function * () {
+export default class RetryableDemoComponent extends Component {
+  @task({
+    enqueue: true,
+    maxConcurrency: 3,
+    evented: true,
+    retryable: backoffPolicy,
+  })
+  *ajaxTask() {
     // simulate slow & unreliable AJAX
     const ms = 2000 + 2000 * Math.random();
     yield timeout(ms);
 
     if (parseInt(ms) % 7 === 0) {
-      throw new Error("Unexpected matrix glitch");
+      throw new Error('Unexpected matrix glitch');
     }
     return {};
-  }).enqueue().maxConcurrency(3).evented().retryable(backoffPolicy),
+  }
 
-  ajaxTaskStarted: on('ajaxTask:started', function(taskInstance) {
+  // Boring demo setup stuff below ...
+
+  onAjaxTaskStarted(taskInstance) {
     const [id] = taskInstance.args;
     this.log(COLORS[id], `Task ${id}: making AJAX request`);
-  }),
+  }
 
-  ajaxTaskRetrying: on('ajaxTask:retrying', function (taskInstance, retryInstance) {
+  onAjaxTaskRetrying(taskInstance, retryInstance) {
     const [id] = taskInstance.args;
     const retryCount = retryInstance.retryCount;
     this.log(COLORS[id], `Task ${id}: retrying request (${retryCount})`);
-  }),
+  }
 
-  ajaxTaskRetried: on('ajaxTask:retried', function (taskInstance, retryInstance) {
+  onAjaxTaskRetried(taskInstance, retryInstance) {
     const [id] = taskInstance.args;
     const retryCount = retryInstance.retryCount;
-    this.log(COLORS[id], `Task ${id}: request succeeded after ${retryCount+1} tries`);
-  }),
+    this.log(
+      COLORS[id],
+      `Task ${id}: request succeeded after ${retryCount + 1} tries`
+    );
+  }
 
-  ajaxTaskSucceeded: on('ajaxTask:succeeded', function(taskInstance) {
+  onAjaxTaskSucceeded(taskInstance) {
     const [id] = taskInstance.args;
     this.log(COLORS[id], `Task ${id}: AJAX done`);
-  }),
+  }
 
-  ajaxTaskErrored: on('ajaxTask:errored', function(taskInstance, error) {
+  onAjaxTaskErrored(taskInstance, error) {
     const [id] = taskInstance.args;
-    this.log(COLORS[id], `Task ${id}: AJAX failed because of '${error.message}'`);
-  }),
+    this.log(
+      COLORS[id],
+      `Task ${id}: AJAX failed because of '${error.message}'`
+    );
+  }
 
-  task0: loopingAjaxTask(0),
-  task1: loopingAjaxTask(1),
-  task2: loopingAjaxTask(2),
-  task3: loopingAjaxTask(3),
-  task4: loopingAjaxTask(4),
+  @task({ on: 'init' }) task0 = loopingAjaxTask(0);
+  @task({ on: 'init' }) task1 = loopingAjaxTask(1);
+  @task({ on: 'init' }) task2 = loopingAjaxTask(2);
+  @task({ on: 'init' }) task3 = loopingAjaxTask(3);
+  @task({ on: 'init' }) task4 = loopingAjaxTask(4);
+
+  @tracked logs = null;
 
   log(color, message) {
     let logs = this.logs || [];
     logs.push({ color, message });
-    this.set('logs', logs.slice(-7));
-  },
+    this.logs = logs.slice(-7);
+  }
 
-  logs: null,
-});
+  trigger(eventName, ...data) {
+    return sendEvent(this, eventName, data);
+  }
+
+  constructor() {
+    super(...arguments);
+
+    addListener(this, 'ajaxTask:started', this, this.onAjaxTaskStarted);
+    addListener(this, 'ajaxTask:retrying', this, this.onAjaxTaskRetrying);
+    addListener(this, 'ajaxTask:retried', this, this.onAjaxTaskRetried);
+    addListener(this, 'ajaxTask:succeeded', this, this.onAjaxTaskSucceeded);
+    addListener(this, 'ajaxTask:errored', this, this.onAjaxTaskErrored);
+
+    registerDestructor(this, () => {
+      removeListener(this, 'ajaxTask:started', this, this.onAjaxTaskStarted);
+      removeListener(this, 'ajaxTask:retrying', this, this.onAjaxTaskRetrying);
+      removeListener(this, 'ajaxTask:retried', this, this.onAjaxTaskRetried);
+      removeListener(
+        this,
+        'ajaxTask:succeeded',
+        this,
+        this.onAjaxTaskSucceeded
+      );
+      removeListener(this, 'ajaxTask:errored', this, this.onAjaxTaskErrored);
+    });
+
+    this.trigger('init');
+  }
+}
 
 function loopingAjaxTask(id) {
-  return task(function * () {
+  return function* () {
     while (true) {
       try {
         yield this.ajaxTask.perform(id);
-      } catch(e) {
+      } catch (e) {
         // Ignoring AJAX failures because we're being naughty.
       }
       yield timeout(2000);
     }
-  }).on('init');
+  };
 }
 // END-SNIPPET