+++ tests/test-helper.js
@@ -1,11 +1,13 @@
 import Application from 'dummy/app';
 import config from 'dummy/config/environment';
+import * as QUnit from 'qunit';
 import { setApplication } from '@ember/test-helpers';
+import { setup } from 'qunit-dom';
 import { start } from 'ember-qunit';
-import defineModifier from 'ember-concurrency-retryable/define-modifier';
-
-defineModifier();
+import 'ember-concurrency-retryable';
 
 setApplication(Application.create(config.APP));
+
+setup(QUnit.assert);
 
 start();