+++ tests/unit/decorator-test.js
@@ -1,16 +1,10 @@
 import { setOwner } from '@ember/application';
 import { module, test } from 'qunit';
-import { task as taskDecorator } from 'ember-concurrency';
-import { task as legacyTaskDecorator } from 'ember-concurrency-decorators';
+import { task } from 'ember-concurrency';
 import { DelayPolicy } from 'ember-concurrency-retryable';
-import { dependencySatisfies } from '@embroider/macros';
 
-const task = dependencySatisfies('ember-concurrency', '^2.0.0-rc.1')
-  ? taskDecorator
-  : legacyTaskDecorator;
-
-module('Unit: decorator use', function() {
-  test("retryable can be used via Task decorator", async function(assert) {
+module('Unit: decorator use', function () {
+  test('retryable can be used via Task decorator', async function (assert) {
     assert.expect(2);
 
     const DELAY_MS = 100;