+++ tests/unit/modules-test.js
@@ -1,24 +1,21 @@
 import { module, test } from 'qunit';
 import {
   retryable,
-  defineModifier,
   Policy,
   DelayPolicy,
   ExponentialBackoffPolicy,
   isEnabled,
   disableTaskRetries,
-  enableTaskRetries
+  enableTaskRetries,
 } from 'ember-concurrency-retryable';
-import PathExportedDefineModifier from 'ember-concurrency-retryable/define-modifier';
 import PathExportedPolicy from 'ember-concurrency-retryable/policies/base';
 import PathExportedDelayPolicy from 'ember-concurrency-retryable/policies/delay';
 import PathExportedExponentialBackoffPolicy from 'ember-concurrency-retryable/policies/exponential-backoff';
 import PathExportedRetryable from 'ember-concurrency-retryable/retryable';
 
-module('Unit: module exports', function() {
-  test("ember-concurrency-retryable", function(assert) {
+module('Unit: module exports', function () {
+  test('ember-concurrency-retryable', function (assert) {
     assert.ok(retryable);
-    assert.ok(defineModifier);
     assert.ok(Policy);
     assert.ok(DelayPolicy);
     assert.ok(ExponentialBackoffPolicy);
@@ -27,15 +24,11 @@ module('Unit: module exports', function() {
     assert.ok(enableTaskRetries);
   });
 
-  test("ember-concurrency-retryable/define-modifier", function(assert) {
-    assert.ok(PathExportedDefineModifier);
-  });
-
-  test("ember-concurrency-retryable/retryable", function(assert) {
+  test('ember-concurrency-retryable/retryable', function (assert) {
     assert.ok(PathExportedRetryable);
   });
 
-  test("ember-concurrency-retryable/policies", function(assert) {
+  test('ember-concurrency-retryable/policies', function (assert) {
     assert.ok(PathExportedPolicy);
     assert.ok(PathExportedDelayPolicy);
     assert.ok(PathExportedExponentialBackoffPolicy);